requirements.txt
annotated-types==0.7.0
anyio==4.12.1
build==1.4.0
CacheControl==0.14.4
certifi==2024.8.30
cffi==2.0.0
charset-normalizer==3.4.4
cleo==2.1.0
crashtest==0.4.1
cryptography==46.0.5
distlib==0.4.0
distro==1.9.0
dotenv==0.9.9
dulwich==1.1.0
fastjsonschema==2.21.2
filelock==3.24.3
findpython==0.7.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
jeepney==0.9.0
jiter==0.13.0
keyring==25.7.0
more-itertools==10.8.0
msgpack==1.1.2
openai==2.26.0
pbs-installer==2026.2.11
pipenv==2026.0.3
pkginfo==1.12.1.2
platformdirs==4.9.2
poetry==2.3.2
poetry-core==2.3.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyproject_hooks==1.2.0
python-dotenv==1.2.2
python-telegram-bot==22.6
RapidFuzz==3.14.3
requests==2.32.5
requests-toolbelt==1.0.0
SecretStorage==3.5.0
shellingham==1.5.4
sniffio==1.3.1
tomlkit==0.14.0
tqdm==4.67.3
trove-classifiers==2026.1.14.14
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
virtualenv==20.38.0
zstandard==0.25.0
